Adding some unicorn dust.
Oh, and a new background image which has some more contrast.
Furthermore, I've added a second planet script that does't add the
authors' avatars to the sidebar. It'll be used on pages with fewer
columns. I've also made all the subpages work by editing index.php and
updating the sub sites. The disclaimer has a white background now, the
"Join Us" page looks a bit fancier and so does the "About Us" site.
I've also updated the stylesheet file to include one more headline type
and some other stuff. Additionally, the sidebar and footer are now
directly inside the subpages which makes it easier to manage them. The
website is pretty much done, the links just have to be checked and the
code needs to pass the W3C test. Apart from that, it'd be nice to make
the website look better on screens that are < 1.036px wide. It can be
done pretty easily by updating the @media screen sections in
css/styles.css - patches are welcome!